requirements.txt
asgiref==3.2.7
Django==3.0.6
pytz==2020.1
sqlparse==0.3.1
django-bootstrap4==1.1.1
django-geojson==3.0.0
soupsieve==2.0
gunicorn==20.0.4
django-leaflet
beautifulsoup4==4.9.0
jsonfield==3.1.0
